+++ .gitignore
@@ -1,6 +1,10 @@
 
 ### WINDARD DEVELOPMENT IGNORES 
 
+case_files/
+
+### MACOS DEFAULT IGNORES
+
 .DS_Store
 
 ### PYTHON GITHUB DEFAULT IGNORES
