+++ test/run_local_test.sh
@@ -1,2 +1,7 @@
 #!/bin/bash
+# pytest --cov=windard test/unit/cost
 pytest --cov=windard test
+
+rm -rf test/unit/layout/problem*_out
+
+#
